Wait for the AWS instance to be fully deleted.
All AWS benchmark jobs are failing at the end of the test (even
though the test is passed), stating that the security group
could not be removed. This happens because we are not waiting
for the VM to be deleted and AWS blocks the deletion of the depending
security group.