Add chaos-pipeline to verify and catch gh#4541.
This new pipeline creates a set of 100 classes in a 2-node
cluster. Performs a backup, deletes the schema and restores
the backup. The issue was that when restoring the data the
second node was out of sync with the first one. This issue
got fixed in 1.24.6, so bumping the WEAVIATE_VERSION too.