Add periodic chaos-pipeline run.
This commit adds a new cron trigger for the chaos-pipeline, so
that it will run on every day basis.
If the trigger is the cron one, then the weaviate version used
is latest, otherwise whatever value passed in input.weaviate_version
or the default.

After running, a slack message will be sent to a webhook to notify
about the failures.