Pass the slack channel in the right format.
When using the slack-github-action, we need to pass the channel
in a hex format, not as a string, otherwise the request fails.
Also, hyperlinks aren't supported by slack workflows, therefore
only the link can be sent as part of the message and adds the
weavite-version-information job in needs, so it gets access to
its context.