Fix replication_tunable_consistenc job.
Fixes #5106

This PR solves couple of problems occuring on the replication
tunable consistency pipeline. The first is that the job was not
executing because of:
./replication_tunable_consistency.sh: line 98: !docker: command not found
However, the pipeline was green. The logic was changed so that
this won't occur again.
And on the other hand, the last step was reading data with replication
level ONE, therefore the read-repair was not kicking it, having data
inconsistency in all the three nodes, which made the job fail.

Also, display Weaviate logs if activate-deactivate multitenant
job fails and update the versions for the raft upgrade journey job.