Rename read-repair pipeline to async-repair.
The pipeline was originally intended to reproduce
https://github.com/weaviate/weaviate/issues/5229,
however it was found out that asynchronous replication
solved the issue (which allowed confirming that if read
repair doesn't kick in, asynchronous replication will).

For that reason the job was submitted in 1.26 with
AsyncEnabled. However, the name used was still read-repair,
what caused confusion.