Add jobs to identify the real Weaviate version under the tag.
This commit removes the weaviate_max_major_version_test_set parameter
and adds a chain of jobs to identify if a job should run or not under
an specific tag. As the tag not always corresponds to the real semver
version, we leverage the weaviate-version-in-image composite action
to obtain the real semver and with few check jobs we can identify on
the fly if the job has the required minimum version to run.