Make the ann-benchmark AWS instance deletion resilient.
It was noted that some of the AWS instances didn't cleanup
properly, incurring in extra costs from our AWS EC2 account.
This commit makes sure that the instance gets deleted and if
for whatever reason it didn't after 5 minutes, the job will
fail. So that the person monitoring the job can take a look
at the AWS' instances or escalate the issue to SRE.