Add rolling/single-pod restarts QPS/latency pipeline.
The purpose of theses pipeline is ensuring that Weaviate
isn't so sensitive to rolling restarts from the point
of view of data ingestion as well as querying.
The pipeline creates one or several collections,
ingest data in the background, then starts a QPS
benchmark which stores latency and QPS and while both
ingestion and QPS is taking place it can either perform:
- a rollout restart
- a single node restart (starting by a non leader and following
by a leader).
The pipeline checks that the QPS and latencies obtained
are within a threshold, also stores relevant metrics and dashboards
for post-analsys.