Add rolling updates QPS/latency pipeline.
The purpose of this pipeline is ensuring that Weaviate
isn't so sensitive to rolling restarts from the point
of view of data ingestion as well as querying.
The pipeline creates one or several collections,
ingest data in the background, then starts a QPS
benchmark which stores latency and QPS and while both
ingestion and QPS is taking place, a rollout restart
of weaviate occurs.
The pipeline checks that the QPS and latencies obtained
are within a threshold.