Add metrics storing in all chaos-pipelines.
This PR allows capturing metrics like CPU usage and memory
usage in the runners in which the chaos pipelines work. For
this, it uses the action
catchpoint/workflow-telemetry-action@master
which acts as a lightweight agent running alongside your workflow steps,
periodically gathering system-level metrics and process-level traces,
and then summarizing them at the end.